Refactor before implementing changes
