+++ src/lib/messages/payloads/CargoMessageSet.ts
@@ -1,9 +1,9 @@
 import * as asn1js from 'asn1js';
 import bufferToArray from 'buffer-to-arraybuffer';
 
-import InvalidMessageError from './InvalidMessageError';
-import Parcel from './Parcel';
-import PayloadPlaintext from './PayloadPlaintext';
+import InvalidMessageError from '../InvalidMessageError';
+import Parcel from '../Parcel';
+import PayloadPlaintext from '../PayloadPlaintext';
 
 /**
  * Plaintext representation of the payload in a cargo message.
